+++ src/NCronJob/Execution/StartupJobManager.cs
@@ -1,3 +1,6 @@
+using System.Globalization;
+using System.Text;
+
 namespace NCronJob;
 
 internal class StartupJobManager(
@@ -8,12 +11,41 @@ internal class StartupJobManager(
     public async Task ProcessStartupJobs(CancellationToken stopToken)
     {
         var startupJobs = jobRegistry.GetAllOneTimeJobs();
-        var startupTasks = startupJobs.Select(definition => CreateExecutionTask(JobRun.Create(observer.Report, definition), stopToken)).ToList();
 
-        if (startupTasks.Count > 0)
+        if (startupJobs.Count == 0)
         {
-            await Task.WhenAll(startupTasks).ConfigureAwait(false);
+            return;
         }
+
+        List<JobRun> jobRuns = [];
+        var startupTasks = startupJobs.Select(definition =>
+        {
+            var jobRun = JobRun.Create(observer.Report, definition);
+            jobRuns.Add(jobRun);
+            return CreateExecutionTask(jobRun, stopToken);
+        });
+
+        await Task.WhenAll(startupTasks).ConfigureAwait(false);
+
+        Exception[] faults = jobRuns
+            .Where(jr => jr.JobDefinition.ShouldCrashOnStartupFailure == true && jr.CurrentState.Type == JobStateType.Faulted)
+            .Select(jr => jr.CurrentState.Fault)
+            .Cast<Exception>()
+            .ToArray();
+
+        if (faults.Length == 0)
+        {
+            return;
+        }
+
+        StringBuilder sb = new();
+        sb.AppendLine("At least one of the startup jobs failed");
+        foreach (var fault in faults)
+        {
+            sb.AppendLine(CultureInfo.InvariantCulture, $"- {fault}");
+        }
+
+        throw new InvalidOperationException(sb.ToString());
     }
 
     private async Task CreateExecutionTask(JobRun job, CancellationToken stopToken) =>